Redesign for AddAllergy
Allergies were originally stored as one tuple per user and a comma delimited string. This has been replaced with a system that stores every individual allergy as a tuple. This is generally better practice and also means that decoding and encoding the allergies from and to comma-delimited strings only happens on receipt and return of information.