Refactor tests to be self sufficient
Originally, certain unit tests relied on other tests to be executed beforehand to work properly. For instance, the TestDeleteEndpoint test relied on the TestSignUpEndpoint test to create a user first. The DeleteEndpoint unit test would then delete the user created in the prior test. These tests have been redesigned to create their test information in the database and clean up any leftover data. This way, a unit test can be executed completely independently and there is no change in the database made after a unit test is complete.