Recommendations reflect user like/dislike
When the recommendations handler fetches a batch of recipes from the API, it will check whether each one has been recommended to the user before and whether or not they have liked/disliked it.

If they have not liked/disliked it, it will be added to the list of recipes to be returned If they have, it will be skipped over. The handler is specified to return 5 recipes. Therefore, if 5 recipes have not been added to the list at the end of processing the batch, the handler will loop again and fetch another batch from the API.